% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_percentual.R
\name{count_percentual}
\alias{count_percentual}
\title{Cria frequências de colunas em relação à decisão}
\usage{
count_percentual(df, ..., decisao)
}
\arguments{
\item{df}{data.frame}

\item{...}{colunas}

\item{decisao}{coluna decisão}
}
\value{
tabela de frequências
}
\description{
Cria frequências de colunas em relação à decisão
}
