% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_bar.R
\name{discrete_bar}
\alias{discrete_bar}
\title{Echarts for discrete scale}
\usage{
discrete_bar(df, grupo, variavel, titulo)
}
\arguments{
\item{df}{Data.frame}

\item{grupo}{Groping variable}

\item{variavel}{discrete items}

\item{titulo}{Title}
}
\value{
Chart that can also be downloaded as png
}
\description{
Echarts for discrete scale
}
\details{
It only charts  2, 3, 5, and 7
levels. It also removes NAs.
}
