% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_funnel.R
\name{gg_funnel}
\alias{gg_funnel}
\title{Creates a ggplot funnel based on a decreasing numeric vector}
\usage{
gg_funnel(x, text = NULL, color = NULL)
}
\arguments{
\item{x}{Decreasing numeric vector}

\item{text}{Text vector to annotate the funnel.}

\item{color}{Vector of colors}
}
\value{
ggplot funnel object
}
\description{
Creates a ggplot funnel based on a decreasing numeric vector
}
\details{
This function needs improvement. I made it ad doc.
}
