% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juris_tree.R
\name{juris_tree}
\alias{juris_tree}
\title{Cria árvore de probabilidades numa distribuição binomial}
\usage{
juris_tree(
  p = NULL,
  probs = list(0.63, c(0.56, 0.78), c(0.89, 0.43, 0.76, 0.25))
)
}
\arguments{
\item{p}{Se você informar uma única probabilidade,
a função admitirá três tentativas de Bernoulli
identicas e independentes, ou seja, uma distribuição
binomial.}

\item{probs}{Lista de probabilidades. Caso queira informar
probabilidades distintas para cada um dos eventos.}
}
\value{
Árvore de probabilidades
}
\description{
Esta função cria uma visualização de árvore de
probabilidades. Ela permite que você mantenha a mesma
probabilidade em cada evento, e.g., qual a probabilidade de
de eu ganhar duas ações de três distribuídas.  Ou calcular
a probabilidade de ganhar uma mesma ação em todas as instâncias,
ex. 1ª, 2ª e superior (STJ, STF), sabendo que para cada
uma dessas instâncias a probabilidade é distinta.
}
