% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_bar.R
\name{jus_bar}
\alias{jus_bar}
\title{Bar plot showing the relationship between the predicted and one predictor variable.}
\usage{
jus_bar(df, group, decision, title)
}
\arguments{
\item{df}{data.frame with predictor and predicted variables.}

\item{group}{predictor.}

\item{decision}{predicted variable.}

\item{title}{plot title.}
}
\value{
bar plot with group in the x axis and decision count in the y axis.
}
\description{
Bar plot showing the relationship between the predicted and one predictor variable.
}
