% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_bench.R
\name{jus_bench}
\alias{jus_bench}
\title{Builds a time series benchmarking processing time.}
\usage{
jus_bench(df, entry_date, matter, decision_date, label,
  peridiocity = "monthly")
}
\arguments{
\item{df}{dataframe with the matter, file date (could be a crime date), and decision date.}

\item{entry_date}{application, loading or crime date}

\item{matter}{matter}

\item{decision_date}{decision or any other reference date.}

\item{label}{label to the matter.}
}
\value{
time series with processing benchmark.
}
\description{
Builds a time series benchmarking processing time.
}
