% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_dt_cjpg.R
\name{jus_dt_cjpg}
\alias{jus_dt_cjpg}
\title{Creates a DT datatable for cjpg}
\usage{
jus_dt_cjpg(df)
}
\arguments{
\item{df}{dataframe ou tibble}
}
\value{
datatable
}
\description{
Creates a DT datatable for cjpg
}
