% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_funnel.R
\name{jus_funnel}
\alias{jus_funnel}
\title{Cria funil echarts a partir dos critérios de exclusão para chegar na base de análise.}
\usage{
jus_funnel(estagios = NULL, subtracoes = NULL, titulo = "")
}
\arguments{
\item{estagios}{vetor com etpas de filtragens}

\item{subtracoes}{vetor com quantidades de processos excluídos}

\item{titulo}{Título do gráfico}
}
\value{
funil echarts.
}
\description{
Cria funil echarts a partir dos critérios de exclusão para chegar na base de análise.
}
