% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_sankey.R
\name{jus_sankey}
\alias{jus_sankey}
\title{Plots echarts sankey diagram from a data.frame}
\usage{
jus_sankey(df)
}
\arguments{
\item{df}{a data.frame of categorical variables}
}
\value{
A sankey ordered by the data.frame column names.
}
\description{
Plots echarts sankey diagram from a data.frame
}
\examples{
\dontrun{
titanic <- as.data.frame(Titanic) \%>\%
           tidyr::uncount(Freq)
jus_sankey(titanic)
}
}
