% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_time.R
\name{jus_time}
\alias{jus_time}
\title{Cria série temporal com tempo de processamento com um ou mais assuntos.}
\usage{
jus_time(base, data_entrada, assunto, data_julgamento, etiqueta)
}
\arguments{
\item{base}{dataframe com assunto, data de entrada ou fato e data do julgamento.}

\item{data_entrada}{data de entrada}

\item{assunto}{assunto}

\item{data_julgamento}{data do julgamento}

\item{etiqueta}{etiqueta para assunto, ex. crimes, temas do consumidor etc.}
}
\value{
Série temporal com highcharts.
}
\description{
Cria série temporal com tempo de processamento com um ou mais assuntos.
}
