% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_comarcas_sp.R
\name{leaflet_comarcas_sp}
\alias{leaflet_comarcas_sp}
\title{Plota mapa interativo das comarcas de São Paulo com número de casos}
\usage{
leaflet_comarcas_sp(df = NULL)
}
\arguments{
\item{df}{dataframe com colunas comarca, n,lat e lng}
}
\value{
mapa interativo
}
\description{
Plota mapa interativo das comarcas de São Paulo com número de casos
}
