% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_bar.R
\name{likert_bar}
\alias{likert_bar}
\title{Echarts for likert scale}
\usage{
likert_bar(df, grupo, variavel, titulo)
}
\arguments{
\item{df}{Data.frame}

\item{grupo}{Groping variable}

\item{variavel}{Likert item}

\item{titulo}{Title}
}
\value{
Chart that can also be downloaded
}
\description{
Echarts for likert scale
}
\details{
It only charts  2, 3, 5, and 7
levels. It also removes NAs.
}
