% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_map.R
\name{mapa}
\alias{mapa}
\title{server}
\usage{
mapa(input, output, session, df)
}
\arguments{
\item{input}{The input object}

\item{output}{The output object}

\item{session}{The session object}

\item{...}{Any additional parameters passed to the module}
}
\description{
server
}
