% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_map.R
\name{mapaui}
\alias{mapaui}
\title{Modulo}
\usage{
mapaui(id)
}
\arguments{
\item{id}{id}

\item{input}{The input object}

\item{output}{The output object}

\item{session}{The session object}

\item{...}{Any additional parameters passed to the module}
}
\description{
This is a page module ...
}
