% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_jus_bench.R
\name{mod_jus_bench_ui}
\alias{mod_jus_bench_ui}
\alias{mod_jus_bench_server}
\title{mod_jus_bench_ui and mod_jus_bench_server}
\usage{
mod_jus_bench_ui(id)

mod_jus_bench_server(input, output, session, timedata)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
