% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_jus_gauge.R
\name{mod_jus_gauge_ui}
\alias{mod_jus_gauge_ui}
\alias{mod_jus_gauge_server}
\title{mod_jus_gauge_ui and mod_jus_gauge_server}
\usage{
mod_jus_gauge_ui(id)

mod_jus_gauge_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
gauge to show predicted results.
}
\keyword{internal}
