% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_this.R
\name{mod_this_ui}
\alias{mod_this_ui}
\alias{mod_this_server}
\title{mod_this_ui and mod_this_server}
\usage{
mod_this_ui(id)

mod_this_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
