% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ToTitleCasePT}
\alias{ToTitleCasePT}
\title{Convertion to title case with lower case for some classes of words.}
\usage{
ToTitleCasePT(string)
}
\arguments{
\item{string}{vector of characters to be converted to title case
but with connector words (one-word prepositions and conjunctions)
and articles (both definite and indefinite) and some pronouns
in lower case.}
}
\value{
vector of characters with the same dimension of \code{string}
}
\description{
In written Portuguese, when converting to title case, it is not usual
to keep in title case some words, like prepositions, conjunctions,
articles and some kinds of pronouns.  This functions locates those
cases and converts them to lower case.
}
\details{
The current list of words converted to lower case is:
\itemize{
\item articles
\itemize{
\item o(s), a(s), um, uma(s), uns
}
\item pronouns
\itemize{
\item me, mim, meu(s), minha(s)
\item te, ti, teu(s), tua(s)
\item lhe(s), seu(s), sua(s)
\item nos, nosso(a)(s)
\item vos, vosso(a)(s)
}
\item prepositions, their contractions and combinations
\itemize{
\item prepositions
\itemize{
\item a, ante, até, após, com, contra, de, desde, em, entre, para,
perante, por, sem, sob, sobre, trás
}
\item contractions
\itemize{
\item à(s), do(a)(s), no(a)(s), pelo(a)(s), pro(a)(s) (informal language)
}
\item combinations
\itemize{
\item ao(s)
}
}
\item conjunctions
\itemize{
\item conforme, conquanto, contudo, durante, embora, enquanto, então,
entretanto, exceto, logo, mas, nem, ou, ora, pois, porém, porque,
porquanto, portanto, quando, quanto, que, se, senão, todavia
}
}

The above list is far from complete or exaustive, mainly due to the absence
of the accidental prepositions and conjuntions, which are words that are not
originally prepositions or conjunctions, but can play those roles in some contexts,
like, for instance \emph{segundo}, which can mean either the numeral \emph{second} or
the prepositional expression \emph{acording to}.
}
