% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_bar.R
\name{jus_bar}
\alias{jus_bar}
\title{Gráfico de barras com variável binária dependente e outra dependente}
\usage{
jus_bar(base, grupo, decisao, titulo)
}
\arguments{
\item{base}{data.frame com dependente e independentes}

\item{grupo}{variável categórica a ocupar eixo x}

\item{decisao}{variável dependente binária}

\item{titulo}{título do gráfico}
}
\value{
gráfico de barras com VI no eixo x e VD no eixo y.
}
\description{
Gráfico de barras com variável binária dependente e outra dependente
}
