% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jus_funnel.R
\name{jus_funnel}
\alias{jus_funnel}
\title{Builds a funnel plot showing criteria and number of dockets removed along the research}
\usage{
jus_funnel(stages = NULL, initial = NULL, removals = NULL,
  title = "")
}
\arguments{
\item{stages}{vector with filtering stages.}

\item{initial}{initial number of dockets.}

\item{removals}{vector with number of dockets removed at each stage.
It should be one element shorter than stages.}

\item{title}{funnel title.}
}
\value{
echarts funnel.
}
\description{
Builds a funnel plot showing criteria and number of dockets removed along the research
}
